Move reviewer to ui.reviewers package
