+++ app/src/main/java/com/appunite/loudius/ui/reviewers/ReviewersViewModel.kt
@@ -9,7 +9,6 @@ import androidx.lifecycle.viewModelScope
 import com.appunite.loudius.common.Screen
 import com.appunite.loudius.common.flatMap
 import com.appunite.loudius.domain.PullRequestRepository
-import com.appunite.loudius.domain.model.Reviewer
 import com.appunite.loudius.network.model.RequestedReviewersResponse
 import com.appunite.loudius.network.model.Review
 import com.appunite.loudius.ui.reviewers.ReviewersSnackbarType.FAILURE
